implemented images api

+++ db/migrations/000004_create_plans_table.up.sql
@@ -2,10 +2,10 @@ CREATE TABLE ekyc_schema.plans (
     id UUID DEFAULT uuid_generate_v4() PRIMARY KEY NOT NULL,
     plan_name ekyc_schema.plan_type NOT NULL,
     is_active BOOLEAN NOT NULL,
-    image_upload_cost INTEGER NOT NULL,
-    face_match_cost INTEGER NOT NULL,
-    ocr_cost INTEGER NOT NULL,
-    daily_base_cost INTEGER NOT NULL,
+    image_upload_cost FLOAT NOT NULL,
+    face_match_cost FLOAT NOT NULL,
+    ocr_cost FLOAT NOT NULL,
+    daily_base_cost FLOAT NOT NULL,
     created_at TIMESTAMP NOT NULL,
     updated_at TIMESTAMP NOT NULL
 );
